+++ dos/src/lib.rs
@@ -1,0 +1,2 @@
+#![allow(clippy::integer_arithmetic)]
+pub mod cli;
